clean code

+++ src/main/java/com/example/springreactivemongodbaggregation/model/InvoiceCountSummary.java
@@ -1,4 +1,3 @@
 package com.example.springreactivemongodbaggregation.model;
 
-public record InvoiceCountSummary(int total, int paid, int canceled, int pending) {
-}
+public record InvoiceCountSummary(double total, double paid, double canceled, double pending) {}
